Update C++ Language usage to C++17
This was an abortive attempt to switch to the Shorter::Namespace::Declaration, however that requires latest, and I'm not sure I want that.

As part of this, I learned that the find_payload can stop being a struct w/ an actual class, and you know, use a bind. Now I understand bind. Now I am happy.