Remove sync writing to disk when queuing
As of this change, the queuing of an item no longer writes it to disk. It's put in a queue to be handled at some other later point in time.

To mitigate testability at this commit, the 'dumb' implementation of draining the queue (PersistAllQueuedItemsToStorage) is added that it can wait on.

Additionally, removing the items from the file system when an ID was removed from the queue is no longer a "thing". The test validating this was removed.