Replace persistence tasks with worker thread
Rather than using a set of tasks & awaits that need to be woken up, this change introduces a worker thread to handle writing the items to disk.
As part of this, it also introduces 'shutting down' state for the worker, so that it can be controlled between running, draining, dropping, and shutdown.

This is intended to work this way:
1. Don't start the worker until initiated by explicit call to EnableQueuingToStorage
2. When the worker starts it checks to see if there are items in the queue.
3. If not, it waits still signalled in someway
4. Copies the items from the queue, and releases the lock
5. Writes the items to disk, checking if we're moving to the drop stage of shutdown
6. Removes the items from the queue that we just wrote to disk.
7. If we're not shutting down, drop the persisted items into the upload queue

Additionally, adds some more traces for clarity when debugging.