Refactor of worker thread in preparation to class-ize
Given the 'worker queue with debounce' is needed in two places, and in reality (well, with a little squinting) they're the same basic queue management, we should factor things out into a slightly more generic process.

This change is the first step by factoring the "work" of the worker into two methods. As part of this, we could no longer just remove the range in the queue, and need to loop through it by hand (boo). This seems to be OK for now.

Additionally:
1. No longer need m_shutdownState to be optional. It's got a "none" state.
2. Since making the PayloadContainer public to allow for the using aliasing, we can remove the find function from being in the header.