Update Pause + Add More Pause-related Tests
Adds a few more pause tests to test specific scenarios when using Pause. As part of this, Pause() itself was updated to use Shutdown() to leverage existing code to terminate background thread etc. This has the slight change that it does cause a "wait" on the pause, but that is expected to drop PDQ