Enable correct debounce behaviour when starting WorkerState
Prior to this change, on starting the worker all items in the queue no
matter the threshold, or idle timeout, would be processed. This means
that any previously queued work would be immediately processed -- potentially
doing work when we didn't intend to have any.

With this change, now on startup the items will be processed in accordance
with any thresholds or idle time outs that might be set. E.g. if you have
a threshold of 4 items, and you queue 5 before calling `.Start()`, upon
calling `.Start()`, those items will be processed.

This includes unit tests to validate this behaviour.

Additionally, there are some light weight refactors in this change to
make things a little clearer & easier to enable the debounce behaviour.