Fixed occasional deadlock / test crash related to worker startup/shutdown
Sometimes, the worker thread would get deadlocked due to the startup & shutdown states overlapping, and various signals on the conditions waiting for the thread to wait up. Specifically, they would wait quickly, and miss anyone else signalling that the item was already woken. This change serializes some of that startup so the initial wait is correctly processed/signaled.

Additional changes:
- Rename m_accessLock to m_itemsLocl
- Fix a trace message from when the code was moved from EventQueue that said "Write To storage worker"
- Fixed up tests to provide the test name as the trace prefix in the worker allowing easier diagnosis.