Support Priority on Background Work -- doesn't wake worker
Rather than have every peice of work potentially wake the thread -- or more accurately _potentially_ wake the thread, allow items to be placed "silently" in the work queue.

This is partially future work to enable queuing of work that otherwise might cause the network stake to wake up -- we don't always want that, since it's expensive.

It also threads this through to the EventQueue itself.