Fix up MixPanelClient to no longer pass unused querystring
When we switched to using post to send the track requests, a "data=" query string turd was left behind on the base URL. Oops.

Additionally, in running this now, it's clear I wasn't quite awake when I last setup the tests -- specifically:
a) Didn't send the request -- it dropped it 'cause of the scaffolding that was actually made to suppor the queuing work
b) even if it had, it didn't _wait_ on it, so the request was killed before it completed the handshake with MixPanel. Doh.

Past me was a muppet.