Removed Concept of Upload From Event Queue
Since really MixpanelClient should be the owner of the upload queue (to avoid creating a multi-level monster of abstraction), we should remove the concept of "upload" from EventQueue. (Which is really a storage queue manager, and will be renamed later)

This removes _almost_ all of the upload concepts, leaving one more for cleanup in a later change.