Removed "TrackSendPriority" concept from API
There's not good reason to provide that direct-send functionality at this point -- this removes it and pushes items into the (non-functioning) queue.

This also removes the IF-def'd test that actually posted a request to the service. It'll come back in some form in the future.

Additional: Changed the JsonObject paramters to IJsonValue 'cause the types in the client are relaxed also.