Added correct priority parameter when adding to the upload queue
When work is processed to disk, it carries a priority with it. This hslps us indicate if we want to actually wake the network stack up to do the upload.

However, since we don't add 1-at-a-time, we need to walk all the items -- if none are normal priority, then we can queue it as low. If there is even one as normal, we have to queue everything as normal.