Refactor timeouts & thresholds so tests can run faster
There were lots of "oh wait for this long", and were _basically_ random values. This wasn't helpful, as it slowed down the execution time of the tests.

So, we expose the right knobs & dials to allow us to crank them down. We also turn off writing to disk, since we don't really need that for most of the tests. You know, except that one... so added a way to toggle them on.

This allows the tests to be cleaned up and shortened.

Additional: The failure batching tests were being a pain in the rear on failures, due to the captured variables in the lambda. So:
1. Simplified it to track counts
2. Turned it into a shared_ptr in the case of exceptions so it doesn't barf on destroyed stack objects.