Adding Suspend / Resuming handling
With the intent of this library to be simply pick it up, an use it merely by providing the key, and instantiating it, and then tracking your events, rather than asking the consumer to hook up many different parts to get to a "functional" model, it means we need to take care of some house keeping.

This means listening to suspend/resume events and pausing (writing any non-uploaded items to disk too), and then starting them again when we're resumed.