Adding Basic Duration Tracker for tracking durations
Because there are cases in the app where we need to track durations, and it's not just timePointA - timePointB differences due to suspension of the app (E.g. wall clock time doesn't cover our needs). we need a helper to track those times.

This is a basic class to track times -- just the boiler plate to get started.