Fixing automation duration to actually work
It wasn't working because I was putting a JsonValue in the property bag -- which we don't support, and wasn't needed.

We did have a unittest to catch it, but since VS 2017 15.6, it's not capturing all the tests if the run fails, and then *fails to report anything* in the test results window if the test process crashes compeletely. So... the test wasn't working, so I didn't catch it.

This change fixes all of that, and now the tests passes.