Added Session Properties
With this, setting a session property will have it attached automatically to the session event sent when the session ends.

Also:
Refactored the super properties to use a return value from InitialzieSuperProperties instead of calling then using the method. They're all just basic "pass through" methods, so I thought this was cleaner.