Changing drain behaviour to post process items
When draining a queue rather than dropping it, we had previously skipped post processing the items. At first, this seems logic, but it's _totally_ not.

Why? Because while the consistency of the specific queue in question is correct, anything that happens in _post processing_ is skipped once the queue is resumed/reloaded. This means, if there is a secondary processing queue that is going to some work with things _after_ they've been processed. Since we weren't calling postprocess, that other peice of work had no idea what was going on. It can't just read the files from disk; it might have to merge them with whatever it's got, which seems to be quite a burden to put on the consumer.

So, instead, we're going to drain it by calling post process, on the assumption that the _post processing_ work is going to be intelligent, and not create more work.