If DropEventsForPrivacy is set before initialization, skip loading data
Loading of any previously persisted items happens on initalization. However, if the customer has said 'Don't send events', we should respect that even after a reinitialization.

This was uncovered with a customer who was actively blocking the Mixpanel endpoints on his router, and saw 100% CPU uage when it failed to upload -- even after he had disabled telemetry. This is a work around for now.