Improve Reliability of a test
The `InitializeAsyncRestoresQueuedToStorageItems` test was quite flakey. After investigation, it looks like it was because of the `SpinWaitForItemCount` was not aware of the atomic nature, causing some flakeyness.

So, changing `SpinWaitForItemCount` to take an atomic, and updating callers help resolve this.

Additionally, cleans up this test to more accurately verify the behaviour