Added optional backoff behaviour during Retries in `BackgroundWorker`
During Retries, it's possible to end up in an loop where you're retrying forever. However, this casues some challenges with bad/failed network situations -- we'll spin forever. This is obviously bad.

As part of resolving that issue, this adds opt-in behaviour in the background worker so that on a _total_ failure of a payload attempt, it will enter a retry state where it will wait before the next retry.

This includes unit tests for the new behaviours.

Additional changes:
a) Test change that wasn't evaluating `shouldKeepProcessing` properly (duh, it's a function)
b) Fixed trace statements to have `: ` infront of them so they're more readable.