Updating the `SendRequestToService` to return a failure reason
As part of trying to resolve spin-waiting when then network fails, we need to understand if it was successful (duh), transiently failed (E.g. service rejected it), or if it was catastrophic (No reason to try again right now, the network is kinda busted)

This change adds that outcome to the result of posting to Mixpanel, allowing the caller to behave slightly differently because of it.