Enable Back off on network queues + auto start when networking is restored
This change enables the back off behaviour in `BackgroundWorker` when consumed by `MixpanelClient` for the profile & track upload workers.

Additionally, it detects when an upload _really_ fails, and adds a listener for when the network is restored to restart the queue.

Additional:
Attempting to fix a flakey test by reducing the timeout