Fixing flakey test in release mode
The `QueueCanBeRestartedAfterRetryLimitIsReachedAndTheQueueIsPaused` test was flakey in release mode; it would fail saying the condition had timed out (When it should have been cleanly signaled).

This was because in release mode the call to `Start()` was triggering the condition before we actually got to wait on it (waiting on a signaled condition will just wait until it's signalled _again_).

So, added a delay to before the condition is signaled. Much better.

Also, while fixing this, tried to change the item threshold after the queue had stopped -- but couldn't. Changed the exception thrown when changing item threshold limits to be when the queue is running not has ever run.