Improved the handling of being unable to write to disk
Since the intent of this component is to get telemetry delivered to Mixpanel, handling this becomes important, otherwise the caller needs to do a bunch of work.

So:
a) If we can't actually write the file to disk for some reason, we just say it's completed and let it moved to uploading. This is 'cause blocking on writing to disk means nothing can make it to the service, which generally seems bad.
b) If we can't create / open the folders we use, lets use temporary, _but also_ not write anything to disk so we don't cross the streams and create other issues.

It's not perfect, but it's a good place to start. If this fixes the issues being observed, we will consider adding a callback to allow reporting of the cases where this is being seen.