Provide fallback handling when stringification of time fails on certain machines
Based on a report from a wonderful customer, it appears that there are certain situations where this call -- when supplied with the Etc/UTC timezone will cause the platform to throw a 'Catastrophic Error' COMException. It's unclear why -- the customer reported it
on Windows 10, 1803 when in the Australian region/timezone.

While unable to repro locally on any of my devices, it was confirmed that using the 'no timezone adjustment' version of the API results in a successful stringification. So, lets try the 'right way' (Since Mixpanel says to supply this information in UTC), and if it fails,
accept that the time could be off by ~12 hours in either direction.

Additional: Tweak the declaration to no longer pass by reference, if only for sanity sake